refactor: break out device and job logic from main file
