+++ src/app/services/createService.ts
@@ -21,7 +21,7 @@ export const createService = (
   name: string,
   sensor: ISensor,
   sendMessage: SendMessage,
-) => {
+): Service => {
   const Service = services[name];
   if (Service == null) {
     throw new Error(`No service for a sensor named '${name}' is known.`);
